pyproject.toml
[project]
name = "pokedex"
version = "0.1.0"
description = ""
authors = [
    {name = "federico",email = "federicogualdi98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "fastapi[standard-no-fastapi-cloud-cli] (>=0.125.0,<0.126.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
]

[tool.poetry]
name = "pokedex"
packages = [{ include = "pokedex", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "debugpy (>=1.8.19,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]


[tool.ruff]
target-version = "py313"
fix = true
show-fixes = true
# max line length
line-length = 120

[tool.ruff.lint]
select = [
    "A",        # flake8-builtins
    "ANN",      # flake8-annotations
    "ARG",      # flake8-unused-arguments
    "ASYNC",    # flake8-async
    "BLE",      # flake8-blind-except
    "C4",       # flake8-comprehensions
    "C90",      # mccabe
    "COM",      # flake8-commas
    "D",        # docstrings
    "E",        # pycodestyle (Error)
    "ERA",      # eradicate
    "F",        # pyflakes/autoflake
    "FLY",      # flynt
    "G",        # flake8-logging-format
    "N",        # pep8-naming
    "I",        # isort
    "PL",       # Pylint
    "PIE",      # flake8-pie
    "PT",       # flake8-pytest-style
    "PYI",      # flake8-pyi
    "Q",        # flake8-quotes
    "RET",      # flake8-return
    "RSE",      # flake8-raise
    "RUF",      # Ruff-specific rules
    "SLF",      # flake8-self
    "SIM",      # flake8-simplify
    "T20",      # flake8-print
    "TD",       # flake8-todos"
    "TRY",      # tryceratops
    "S",        # flake8-bandit
    "UP",       # pyupgrade
    "W",        # pycodestyle (Warning)
    "YTT",      # flake8-2020
]
ignore = [
    "ANN002",   # Missing type annotation for *args
    "ANN003",   # Missing type annotation for **kwargs
    "G004",     # Logging statement uses f-string
    "TD002",    # Missing author in TODO; try: # TODO(<author_name>): ...
    "TD003",    # Missing issue link on the line following this TODO
    "TRY003",   # Avoid specifying long messages outside the exception class
    "D407",     # Missing dashed underline after section
    "D406",     # Section name should end with a newline
    "UP007",    # Use `X | Y` for type annotations
]
fixable = [
    "C4",       # flake8-comprehensions
    "COM",      # flake8-commas
    "I",        # isort
    "Q",        # flake8-quotes
    "F401",     # Unused imports
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["ARG001", "ARG002", "PLR0913", "PLR2004", "S101", "SLF001", "C901"]

[tool.ruff.lint.pydocstyle]
# Use Google-style docstrings.
convention = "google"

[tool.ruff.lint.isort]
# Don't force all from imports to appear on their own line.
force-single-line = true
# Use a single line between direct and from import
lines-between-types = 1

[tool.ruff.lint.flake8-annotations]
suppress-none-returning = true

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.lint.pylint]
max-args = 8
